// Loop until all numbers in sequence visited or found that it is impossible         // to visit all the numbers in sequence         while(nextNum <= matrix.length * matrix[0].length)         {             // Reset             canGoLeft = canGoRight = canGoDown = canGoUp = true;              // Determine which directions are possible             if(row - 1 < 0)                 canGoUp= false;             if(row + 1 == matrix.length)                 canGoDown = false;             if(col - 1 < 0)                 canGoLeft = false;             if(col + 1 == matrix[0].length)                 canGoRight = false;              // Check up for next number in sequence             if(canGoUp && nextNum == matrix[row - 1][col])             {                 nextNum++;                 row--;             }             // Check down for next number in sequence             else if(canGoDown && nextNum == matrix[row + 1][col])             {                 nextNum++;                 row++;             }             // Check left for next number in sequence             else if(canGoLeft && nextNum == matrix[row][col - 1])             {                 nextNum++;                 col--;             }             // Check right for next number in sequence             else if(canGoRight && nextNum == matrix[row][col + 1])             {                 nextNum++;                 col++;             }             // Failed             else             {                 return false;             }         }          return true;     }
You are given a n x m matrix, which contains all the integers from 1 to n * m, inclusive, each exactly once.

Initially you are standing in the cell containing the number 1. On each turn you are allowed to move to an adjacent cell, i.e. to a cell that shares a common side with the one you are standing on now. It is prohibited to visit any cell more than once.

Check if it is possible to visit all the cells of the matrix in the order of increasing numbers in the cells, i.e. get to the cell with the number 2 on the first turn, then move to 3, etc.

Example

For

matrix = [[1, 4, 5], 
          [2, 3, 6]]
the output should be
findPath(matrix) = true;

For

matrix = [[1, 3, 6], 
          [2, 4, 5]]
the output should be
findPath(matrix) = false.

Input/Output

[time limit] 3000ms (java)
[input] array.array.integer matrix

A non-empty 2-dimensional array of integers representing a rectangular matrix.

Constraints:
1 ≤ matrix.length ≤ 5,
1 ≤ matrix[0].length ≤ 10,
1 ≤ matrix[i][j] ≤ 15.

[output] boolean